pegc: extract static analyses into src/pegc/analysis.rs
Pure refactor. Moves `analyze_left_recursion`, `compute_nullable`,
`pattern_nullable`, `compute_first_calls`, `collect_first_calls`, and
`tarjan_sccs` out of `compiler.rs` to make room for the FIRST/FOLLOW
computations introduced in the follow-up. No behaviour change.
